+++ CHANGELOG.md
@@ -24,6 +24,10 @@
   [Ben Yohay](https://github.com/byohay)
   [#861](https://github.com/CocoaPods/Xcodeproj/pull/861)
 
+* Add `code_coverage_targets` support for `TestAction`.  
+  [Joey Dong](https://github.com/joeydong)
+  [#874](https://github.com/CocoaPods/Xcodeproj/pull/874)
+
 ##### Bug Fixes
 
 * Fix undefined method 'downcase' for `nil`